+++ apps/client/src/components/admin/manage/cad-settings/api-token-tab.tsx
@@ -5,11 +5,11 @@ import { Form, Formik, FormikHelpers } from "formik";
 import useFetch from "lib/useFetch";
 import { useTranslations } from "use-intl";
 import { SettingsFormField } from "components/form/SettingsFormField";
-import { SettingsTabs } from "src/pages/admin/manage/cad-settings";
 import { toastMessage } from "lib/toastMessage";
 import type { DeleteCADApiTokenData, PutCADApiTokenData } from "@snailycad/types/api";
 import { getAPIUrl } from "@snailycad/utils/api-url";
 import { TabsContent } from "@radix-ui/react-tabs";
+import { SettingsTabs } from "components/admin/cad-settings/layout";
 
 export function ApiTokenTab() {
   const common = useTranslations("Common");
@@ -146,7 +146,7 @@ export function ApiTokenTab() {
               />
             </SettingsFormField>
 
-            <div className="flex">
+            <div className="flex justify-end">
               {cad?.apiTokenId ? (
                 <Button
                   onPress={() => handleRegenerate(setFieldValue)}
