+++ apps/client/src/components/admin/manage/cad-settings/general/cad-features-tab.tsx
@@ -12,13 +12,13 @@ import {
 } from "@snailycad/types";
 import { Button, Loader, TextField, SelectField, SwitchField } from "@snailycad/ui";
 import { SettingsFormField } from "components/form/SettingsFormField";
-import { SettingsTabs } from "src/pages/admin/manage/cad-settings";
 import { toastMessage } from "lib/toastMessage";
 import { DEFAULT_DISABLED_FEATURES, DEFAULT_FEATURE_OPTIONS } from "hooks/useFeatureEnabled";
 import type { PutCADFeaturesData } from "@snailycad/types/api";
 import Link from "next/link";
 import { BoxArrowUpRight } from "react-bootstrap-icons";
 import { TabsContent } from "@radix-ui/react-tabs";
+import { SettingsTabs } from "components/admin/cad-settings/layout";
 
 const featuresWithURL: string[] = [
   Feature.BLEETER,
