+++ apps/client/src/components/admin/manage/cad-settings/misc-features/license-number-tab.tsx
@@ -7,7 +7,7 @@ import { useAuth } from "context/AuthContext";
 import { Form, Formik } from "formik";
 import useFetch from "lib/useFetch";
 import { useTranslations } from "use-intl";
-import { cleanValues } from "../misc-settings-tab";
+import { cleanValues } from "./other-misc-tab";
 import { toastMessage } from "lib/toastMessage";
 import { MiscCadSettings } from "@snailycad/types";
