+++ packages/schemas/src/admin/index.ts
@@ -4,12 +4,12 @@ export const CAD_SETTINGS_SCHEMA = z.object({
   name: z.string().min(2).max(255),
   areaOfPlay: z.string().max(255),
   steamApiKey: z.string().max(255),
-  towWhitelisted: z.boolean(),
-  taxiWhitelisted: z.boolean(),
-  whitelisted: z.boolean(),
-  roleplayEnabled: z.boolean(),
-  registrationCode: z.string().max(255),
-  businessWhitelisted: z.boolean(),
+  towWhitelisted: z.boolean().optional(),
+  taxiWhitelisted: z.boolean().optional(),
+  whitelisted: z.boolean().optional(),
+  roleplayEnabled: z.boolean().optional(),
+  registrationCode: z.string().max(255).optional(),
+  businessWhitelisted: z.boolean().optional(),
   image: z.any().nullish(),
 });
